Find repository only for changed dependencies
