+++ server/loader/versions/common.ts
@@ -1,13 +1,9 @@
-import { readFileSync } from 'node:fs'
-import { createRequire } from 'node:module'
-
 import type {
   Dependency,
   DependencyName,
   DependencyVersion,
   FilePath,
-  LoadedFile,
-  Repository
+  LoadedFile
 } from '../../../common/types.ts'
 
 /**
@@ -54,85 +50,4 @@ export function splitPackage(pkg: string): {
     }
   }
   return { name: pkg as DependencyName, version: '' as DependencyVersion }
-}
-
-interface PackageJson {
-  bugs?: { url: string } | string
-  homepage?: string
-  private?: true
-  repository?: { type?: string; url: string } | string
-}
-
-function extractRepositoryUrl(packageJson: string): null | string {
-  let pkg = JSON.parse(packageJson) as PackageJson
-  if (pkg.private) return null
-  if (pkg.repository) {
-    if (typeof pkg.repository === 'string') {
-      return normalizeRepositoryUrl(pkg.repository)
-    }
-    if (typeof pkg.repository === 'object' && pkg.repository.url) {
-      return normalizeRepositoryUrl(pkg.repository.url)
-    }
-  }
-  if (pkg.homepage?.includes('github.com')) {
-    return normalizeRepositoryUrl(pkg.homepage)
-  }
-  if (typeof pkg.bugs === 'object' && pkg.bugs.url.includes('github.com')) {
-    return normalizeRepositoryUrl(pkg.bugs.url)
-  }
-  return null
-}
-
-const GIT_URL_POSTFIX = /\.git(#\w+)?$/
-
-function normalizeRepositoryUrl(url: string): Repository {
-  if (url.startsWith('git+http')) {
-    url = url.replace('git+', '').replace(GIT_URL_POSTFIX, '')
-  }
-  if (url.startsWith('git+ssh')) {
-    url = url.replace('git+ssh://git@', 'https://').replace(GIT_URL_POSTFIX, '')
-  }
-  if (url.startsWith('github:')) {
-    url = url.replace('github:', 'https://github.com/')
-  }
-  if (url.includes('github.com')) {
-    url = url.replace(/\.git$/, '')
-    url = url.replace(/#.*$/, '')
-    url = url.replace(/\/issues.*$/, '')
-    url = url.replace(/\/tree.*$/, '')
-    if (!url.startsWith('http')) {
-      let match = url.match(/github\.com[/:](.+)/)
-      if (match) {
-        url = `https://github.com/${match[1]}`
-      }
-    }
-  } else if (
-    !url.includes('://') &&
-    url.includes('/') &&
-    !url.startsWith('.')
-  ) {
-    url = `https://github.com/${url}`
-  }
-  return url as Repository
-}
-
-export function findRepositorySource(
-  url: object | string | undefined,
-  name: string
-): Repository {
-  if (typeof url === 'string' && url.startsWith('git')) {
-    return normalizeRepositoryUrl(url)
-  } else {
-    try {
-      let require = createRequire(import.meta.url)
-      let repoUrl = extractRepositoryUrl(
-        readFileSync(require.resolve(`${name}/package.json`), 'utf8')
-      )
-      if (repoUrl) return repoUrl as Repository
-    } catch {
-      // Package not found in node_modules
-    }
-
-    return `https://www.npmjs.com/package/${name}` as Repository
-  }
 }
