+++ server/loader/versions/github-actions.ts
@@ -116,7 +116,6 @@ export const githubActions = {
           let dep = dependency({
             from: 'github-actions',
             name: action.name,
-            repository: `https://github.com/${action.name}`,
             source: file.path,
             type: 'github-actions',
             version: action.version
